requirements.txt
Django==2.1.7
django-querycount==0.7.0
factory-boy==2.11.1
Faker==1.0.2
psycopg2==2.7.7
psycopg2-binary==2.7.7
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
text-unidecode==1.2
